frontend: CreateResourceForm: Refactor structure for create resource form logic
